# lib/api-zod/src/generated/types/guild.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guilds Onchain Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Guild {
  id: number;
  name: string;
  description?: string;
  collectionAddress: string;
  memberCount: number;
  strength: number;
  rank: number;
  ownerWallet: string;
  createdAt: string;
}