# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guilds Onchain Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  guildId: number;
  guildName: string;
  memberCount: number;
  strength: number;
  weeklyPoints: number;
}