# lib/api-zod/src/generated/types/member.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guilds Onchain Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberRole } from "./memberRole";

export interface Member {
  id: number;
  walletAddress: string;
  discordHandle?: string;
  twitterHandle?: string;
  guildId?: number;
  role: MemberRole;
  points: number;
  createdAt: string;
}