# lib/api-zod/src/generated/types/registerGuildRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guilds Onchain Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterGuildRequest {
  name: string;
  description?: string;
  collectionAddress: string;
  ownerWallet: string;
  collectionName?: string;
  website?: string;
  twitter?: string;
  discord?: string;
}