# lib/api-zod/src/generated/types/registerMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guilds Onchain Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterMemberRequest {
  walletAddress: string;
  discordHandle?: string;
  twitterHandle?: string;
  referralCode?: string;
}